fix: moved callbacks out of app.py into callbacks.py
